Added DB reset with dialog box

+++ app/src/main/java/ru/vysokov/phonebook/SettingsActivity.java
@@ -1,12 +1,19 @@
 package ru.vysokov.phonebook;
 
 import androidx.appcompat.app.AppCompatActivity;
+
+import android.app.AlertDialog;
+import android.app.Dialog;
+import android.content.DialogInterface;
 import android.content.Intent;
+import android.content.res.Resources;
 import android.database.sqlite.SQLiteDatabase;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
 import android.widget.ImageView;
+import android.widget.TextView;
+import android.widget.Toast;
 
 import java.util.LinkedList;
 
@@ -18,9 +25,17 @@ public class SettingsActivity extends AppCompatActivity {
         setContentView(R.layout.activity_settings);
 
         Button buttonClearDB = findViewById(R.id.buttonClearDB);
-        buttonClearDB.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
+
+        ImageView imageBack = (ImageView) findViewById(R.id.back);
+        Intent mainActivity = new Intent(this, MainActivity.class);
+        imageBack.setOnClickListener(view -> startActivity(mainActivity));
+
+        buttonClearDB.setOnClickListener(view -> {
+            AlertDialog.Builder abuilder = new AlertDialog.Builder(SettingsActivity.this);
+            abuilder.setTitle("Сбросить все данные");
+            abuilder.setMessage("Вы уверены?");
+
+            abuilder.setPositiveButton("Да", (dialog, which) -> {
                 SQLiteDatabase database = new DBHelper(getApplicationContext()).getWritableDatabase();
                 database.execSQL("DELETE FROM individual");
                 database.execSQL("DELETE FROM company");
@@ -29,19 +44,15 @@ public class SettingsActivity extends AppCompatActivity {
                 LinkedList userList = MainActivity.getUsers();
                 userList.clear();
                 MainActivity.setUsers(userList);
+                dialog.dismiss();
 
-            }
+                Toast toast = Toast.makeText(getApplicationContext(), "Данные успешно удалены", Toast.LENGTH_SHORT);
+                toast.show();
+            });
+
+            abuilder.setNegativeButton("Нет", (dialog, which) -> dialog.cancel());
+            AlertDialog alert = abuilder.create();
+            alert.show();
         });
-
-        ImageView imageBack = (ImageView) findViewById(R.id.back);
-        Intent mainActivity = new Intent(this, MainActivity.class);
-        imageBack.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                startActivity(mainActivity);
-            }
-        });
-
     }
-
 }
